@@ -10,42 +10,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.1.3",
-    "@angular/cdk": "17.1.2",
-    "@angular/common": "17.1.3",
-    "@angular/compiler": "17.1.3",
-    "@angular/core": "17.1.3",
-    "@angular/forms": "17.1.3",
-    "@angular/platform-browser": "17.1.3",
-    "@angular/platform-browser-dynamic": "17.1.3",
-    "@angular/router": "17.1.3",
+    "@angular/animations": "19.0.7",
+    "@angular/cdk": "19.0.5",
+    "@angular/common": "19.0.7",
+    "@angular/compiler": "19.0.7",
+    "@angular/core": "19.0.7",
+    "@angular/forms": "19.0.7",
+    "@angular/platform-browser": "19.0.7",
+    "@angular/platform-browser-dynamic": "19.0.7",
+    "@angular/router": "19.0.7",
     "@fontsource-variable/inter": "^5.0.17",
     "@ng-icons/core": "^27.2.0",
     "@ng-icons/lucide": "^27.2.0",
-    "@ngrx/effects": "^17.1.1",
-    "@ngrx/entity": "^17.1.1",
-    "@ngrx/store": "^17.1.1",
-    "@ngrx/store-devtools": "^17.1.1",
+    "@ngrx/effects": "19.0.0",
+    "@ngrx/entity": "19.0.0",
+    "@ngrx/operators": "^18.0.0",
+    "@ngrx/store": "19.0.0",
+    "@ngrx/store-devtools": "19.0.0",
     "bootstrap": "^5.3.2",
     "font-awesome": "^4.7.0",
     "https-proxy-agent": "^7.0.4",
     "idb": "^8.0.0",
     "jquery": "^3.7.1",
+    "ngx-sonner": "^3.0.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "xml2js": "^0.6.2",
-    "zone.js": "0.14.4"
+    "zone.js": "0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.1.4",
-    "@angular-devkit/core": "17.1.4",
-    "@angular-devkit/schematics": "17.1.4",
-    "@angular/cli": "~17.1.0",
-    "@angular/compiler-cli": "17.1.3",
-    "@ngrx/schematics": "^17.1.1",
-    "@nx/angular": "18.0.7",
-    "@nx/workspace": "18.0.7",
-    "@schematics/angular": "17.1.4",
+    "@angular-devkit/build-angular": "19.0.7",
+    "@angular-devkit/core": "19.0.7",
+    "@angular-devkit/schematics": "19.0.7",
+    "@angular/cli": "~19.0.0",
+    "@angular/compiler-cli": "19.0.7",
+    "@ngrx/schematics": "19.0.0",
+    "@nx/angular": "20.3.2",
+    "@nx/workspace": "20.3.2",
+    "@schematics/angular": "19.0.7",
     "@types/jasmine": "~4.3.0",
     "@types/xml2js": "^0.4.13",
     "autoprefixer": "^10.4.0",
@@ -55,13 +57,13 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "nx": "18.0.7",
+    "nx": "20.3.2",
     "postcss": "^8.4.5",
     "prettier": "^3.4.2",
     "prettier-plugin-organize-attributes": "^1.0.0",
     "prettier-plugin-organize-imports": "^3.2.4",
     "prettier-plugin-tailwindcss": "^0.5.14",
     "tailwindcss": "^3.0.2",
-    "typescript": "5.3.3"
+    "typescript": "5.6.3"
   }
 }
