@@ -1,16 +1,14 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
+import { NgxSonnerToaster } from 'ngx-sonner';
 
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
     styleUrls: ['./app.component.css'],
     standalone: true,
-    imports: [RouterOutlet]
+    imports: [RouterOutlet, NgxSonnerToaster]
 })
 export class AppComponent {
-  title = 'VectorSearchUI';
-  constructor() {
-  }
 
 }
