@@ -2,11 +2,10 @@ import { Component } from '@angular/core';
 import { HeaderComponent } from '../shared/header/header.component';
 
 @Component({
-    selector: 'app-home',
-    templateUrl: './home.component.html',
-    styleUrls: ['./home.component.css'],
-    standalone: true,
-    imports: [HeaderComponent]
+  selector: 'app-home',
+  templateUrl: './home.component.html',
+  styleUrls: ['./home.component.css'],
+  imports: [HeaderComponent],
 })
 export class HomeComponent {
   title: string = 'Home';
