@@ -1,19 +1,18 @@
 export interface IHighlightingResult {
-    autnresponse: Autnresponse
-  }
-  
-  export interface Autnresponse {
-    action: string
-    response: string
-    responsedata: Responsedata
-  }
-  
-  export interface Responsedata {
-    hit: Hit
-  }
-  
-  export interface Hit {
-    highlighted: string
-    content: string
-  }
-  
+  autnresponse: Autnresponse;
+}
+
+export interface Autnresponse {
+  action: string;
+  response: string;
+  responsedata: Responsedata;
+}
+
+export interface Responsedata {
+  hit: Hit;
+}
+
+export interface Hit {
+  highlighted: string;
+  content: string;
+}
