@@ -1,22 +1,21 @@
 export interface LanguageIdentification {
-    autnresponse: Autnresponse
-  }
-  
-  export interface Autnresponse {
-    action: string
-    response: string
-    responsedata: Responsedata
-  }
-  
-  export interface Responsedata {
-    language: string
-    languageiso: string
-    languageencoding: string
-    languagedirection: string
-    languagescripts: Languagescripts
-  }
-  
-  export interface Languagescripts {
-    unicodeblock: string
-  }
-  
+  autnresponse: Autnresponse;
+}
+
+export interface Autnresponse {
+  action: string;
+  response: string;
+  responsedata: Responsedata;
+}
+
+export interface Responsedata {
+  language: string;
+  languageiso: string;
+  languageencoding: string;
+  languagedirection: string;
+  languagescripts: Languagescripts;
+}
+
+export interface Languagescripts {
+  unicodeblock: string;
+}
