@@ -1,6 +1,6 @@
 export interface ISearchResultItem {
-    title:string;
-    summary:string;
-    reference:string;
-    autnrank:string;
+  title: string;
+  summary: string;
+  reference: string;
+  autnrank: string;
 }
