@@ -1,7 +1,7 @@
 export interface IResultSummary {
-    numhits: number;
-    predicted: string;
-    totaldbdocs: number;
-    totaldbsecs: number;
-    totalhits: number;
+  numhits: number;
+  predicted: string;
+  totaldbdocs: number;
+  totaldbsecs: number;
+  totalhits: number;
 }
