@@ -1,16 +1,15 @@
-import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { NgFor } from '@angular/common';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 
 @Component({
-    selector: 'app-dynamic-valid-choices',
-    templateUrl: './dynamic-valid-choices.component.html',
-    styleUrls: ['./dynamic-valid-choices.component.css'],
-    standalone: true,
-    imports: [NgFor]
+  selector: 'app-dynamic-valid-choices',
+  templateUrl: './dynamic-valid-choices.component.html',
+  styleUrls: ['./dynamic-valid-choices.component.css'],
+  imports: [NgFor],
 })
 export class DynamicValidChoicesComponent {
   @Input() validChoices: string[] = [];
-  
+
   @Output() choiceSelected = new EventEmitter<string>();
 
   selectChoice(choice: string) {
