@@ -1,26 +1,26 @@
 import { Component, computed, HostListener, inject, Input, signal } from '@angular/core';
-import { ISearchResultItem } from 'src/app/interfaces/IsearchResultItem';
-import { SearchResultItemComponent } from '../search-result-item/search-result-item.component';
-import { NgIf, NgFor } from '@angular/common';
-import { environment } from 'src/environments/environment';
 import { DomSanitizer } from '@angular/platform-browser';
-import { provideIcons, NgIcon } from '@ng-icons/core';
+import { NgIcon, provideIcons } from '@ng-icons/core';
 import { lucideX } from '@ng-icons/lucide';
-import { SettingsService } from 'src/app/services/settings.service';
 import { Subscription } from 'rxjs';
+import { ISearchResultItem } from 'src/app/interfaces/IsearchResultItem';
+import { SettingsService } from 'src/app/services/settings.service';
+import { environment } from 'src/environments/environment';
+import { SearchResultItemComponent } from '../search-result-item/search-result-item.component';
 
 @Component({
-    selector: 'app-search-results',
-    templateUrl: './search-results.component.html',
-    styleUrls: ['./search-results.component.css'],
-    standalone: true,
-    imports: [SearchResultItemComponent, NgIcon],
-    viewProviders: [provideIcons({ lucideX })]
+  selector: 'app-search-results',
+  templateUrl: './search-results.component.html',
+  styleUrls: ['./search-results.component.css'],
+  imports: [SearchResultItemComponent, NgIcon],
+  viewProviders: [provideIcons({ lucideX })],
 })
 export class SearchResultsComponent {
-
   enableView: boolean = localStorage.getItem('viewEnabled') === 'true';
-  constructor(private domSanitizer: DomSanitizer, private settingsService: SettingsService) {
+  constructor(
+    private domSanitizer: DomSanitizer,
+    private settingsService: SettingsService
+  ) {
     this.sanitizer = domSanitizer;
   }
   private readonly sanitizer = inject(DomSanitizer);
@@ -28,7 +28,7 @@ export class SearchResultsComponent {
   @Input() currentPage = 1;
   @Input() totalPages = 1;
   @Input() searchResults: ISearchResultItem[] = [];
-  
+
   canClose = false;
   timeout: any;
   readonly previewItem = signal<ISearchResultItem | null>(null);
@@ -41,24 +41,25 @@ export class SearchResultsComponent {
     }
     let params = new URLSearchParams();
     params.append('reference', item.reference);
-    params.append('NoACI','true');
+    params.append('NoACI', 'true');
     //params.append('SecurityInfo', encodeURIComponent(localStorage.getItem('token') as string));
-    params.append('EmbedImages','true');  
-    params.append('StripScript','true');
-    params.append('Links','');
-    params.append('OriginalBaseURL','true');
-    params.append('Boolean','true');
-    params.append('OutputType','HTML');
-    return this.sanitizer.bypassSecurityTrustResourceUrl(`${environment.view_api}?Action=View&${params.toString()}`);
+    params.append('EmbedImages', 'true');
+    params.append('StripScript', 'true');
+    params.append('Links', '');
+    params.append('OriginalBaseURL', 'true');
+    params.append('Boolean', 'true');
+    params.append('OutputType', 'HTML');
+    return this.sanitizer.bypassSecurityTrustResourceUrl(
+      `${environment.view_api}?Action=View&${params.toString()}`
+    );
   });
 
-  loadPage(page: number) {
-  }
-  
+  loadPage(page: number) {}
+
   showPreview(item: ISearchResultItem): void {
     this.previewItem.set(item);
     clearTimeout(this.timeout);
-    this.timeout = setTimeout(() => this.canClose = true, 100);
+    this.timeout = setTimeout(() => (this.canClose = true), 100);
   }
 
   @HostListener('document:keydown.escape')
@@ -76,16 +77,16 @@ export class SearchResultsComponent {
 
     // if we click outside the preview, close it
     if (!(event.target as HTMLElement).closest('.preview')) {
-      this.hidePreview()
+      this.hidePreview();
     }
-  }  
+  }
 
-  ngOnInit() {  
-    this.subscription = this.settingsService.previewEnabled$.subscribe((enabled) => {
+  ngOnInit() {
+    this.subscription = this.settingsService.previewEnabled$.subscribe(enabled => {
       this.enableView = enabled;
     });
   }
-  ngOnDestroy() { 
+  ngOnDestroy() {
     this.subscription.unsubscribe();
   }
 
