@@ -1,17 +1,14 @@
-import { HttpClient } from '@angular/common/http';
 import { Component } from '@angular/core';
 import { AnswerBankService } from '../services/answer-bank.service';
 
 @Component({
   selector: 'app-people-also-asked',
-  standalone: true,
   imports: [],
   templateUrl: './people-also-asked.component.html',
-  styleUrl: './people-also-asked.component.css'
+  styleUrl: './people-also-asked.component.css',
 })
 export class PeopleAlsoAskedComponent {
-
-  constructor(private answerbanksvc:AnswerBankService) { }
+  constructor(private answerbanksvc: AnswerBankService) {}
   questions: any[] = []; // Holds the list of questions
 
   ngOnInit() {
@@ -19,12 +16,12 @@ export class PeopleAlsoAskedComponent {
   }
 
   loadInitialQuestions() {
-    this.answerbanksvc.getrelatedquestions().subscribe((data) => {
+    this.answerbanksvc.getrelatedquestions().subscribe(data => {
       this.questions = data.map((q: any) => ({
         ...q,
         expanded: false,
         loading: false,
-        answer: null
+        answer: null,
       }));
     });
   }
@@ -43,7 +40,7 @@ export class PeopleAlsoAskedComponent {
   loadAnswer(question: any) {
     question.loading = true;
     this.paaService.getAnswer(question.id).subscribe(
-      (answer) => {
+      answer => {
         question.answer = answer;
         question.loading = false;
       },
