@@ -3,10 +3,9 @@ import { BehaviorSubject } from 'rxjs';
 import { Answer } from '../interfaces/IAnswerServerResponse';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class QuestionService {
-
   private questionSubject = new BehaviorSubject<string>('');
   private answerSubject = new BehaviorSubject<Answer[]>([]);
 
