@@ -1,14 +1,14 @@
-import { CanActivateFn, Router } from '@angular/router';
 import { inject } from '@angular/core';
+import { CanActivateFn, Router } from '@angular/router';
 import { LoginService } from '../../services/login.service';
 
-export const authGuard: CanActivateFn =  (route, state) => {
-    const router = inject(Router);
-    const isAuthenticated = inject(LoginService).isAuthenticated;
+export const authGuard: CanActivateFn = (route, state) => {
+  const router = inject(Router);
+  const isAuthenticated = inject(LoginService).isAuthenticated;
 
-    if (!isAuthenticated) {
-        return router.createUrlTree(['/login']);
-    }
+  if (!isAuthenticated) {
+    return router.createUrlTree(['/login']);
+  }
 
-    return true;
+  return true;
 };
