@@ -1,35 +1,31 @@
 import { AsyncPipe, NgClass } from '@angular/common';
 import { Component, EventEmitter, Output, inject } from '@angular/core';
+import { FormsModule } from '@angular/forms';
 import { NgIcon } from '@ng-icons/core';
 import { Store } from '@ngrx/store';
+import { System } from 'src/app/interfaces/IAnswerServerGetStatusResponse';
+import { AnswerService } from 'src/app/services/answer.service';
+import { SettingsService } from 'src/app/services/settings.service';
 import { SettingsActions } from '../../../state/settings/settings.actions';
 import {
   selectShowIdolSearchResults,
   selectShowVectorSearchResults,
 } from '../../../state/settings/settings.selectors';
-import { DahService } from 'src/app/services/dah.service';
-import { Database } from 'src/app/interfaces/IDahResponse';
-import { FormsModule } from '@angular/forms';
-import { AnswerService } from 'src/app/services/answer.service';
-import { System } from 'src/app/interfaces/IAnswerServerGetStatusResponse';
-import { SettingsService } from 'src/app/services/settings.service';
 
 @Component({
   selector: 'app-settings-dialog',
   templateUrl: './settings-dialog.component.html',
   styleUrls: ['./settings-dialog.component.css'],
-  standalone: true,
   imports: [NgIcon, AsyncPipe, NgClass, FormsModule],
 })
 export class SettingsDialogComponent {
-
   private readonly store = inject(Store);
   private readonly answerService = inject(AnswerService);
   readonly showVectorSearchResults$ = this.store.select(selectShowVectorSearchResults);
   readonly showIdolSearchResults$ = this.store.select(selectShowIdolSearchResults);
   isComparisonSearchDisabled: boolean = true;
-  constructor(private settingsService: SettingsService) { 
-    this.settingsService.previewEnabled$.subscribe((enabled) => {
+  constructor(private settingsService: SettingsService) {
+    this.settingsService.previewEnabled$.subscribe(enabled => {
       this.viewEnabled = enabled;
     });
   }
@@ -41,26 +37,52 @@ export class SettingsDialogComponent {
   querySystems: QuerySystem[] = ['Dictionary', 'Index', 'Answerbank'];
   selectedQms?: QuerySystem;
 
-  defaultOperators: DefaultOperator[] = ['WNEAR', 'DNEAR' , 'YNEAR' , 'NNEAR' , 'XNEAR' , 'AND' , 'OR' , 'BEFORE' , 'AFTER' , 'SENTENCE' , 'DSENTENCE' , 'PARAGRAPH' ];
+  defaultOperators: DefaultOperator[] = [
+    'WNEAR',
+    'DNEAR',
+    'YNEAR',
+    'NNEAR',
+    'XNEAR',
+    'AND',
+    'OR',
+    'BEFORE',
+    'AFTER',
+    'SENTENCE',
+    'DSENTENCE',
+    'PARAGRAPH',
+  ];
   selectedOperator?: DefaultOperator;
 
-  summaryOptions: SummaryOptions[] = ['Concept', 'Context', 'Quick', 'ParagraphConcept', 'ParagraphContext', 'RAG', 'Vector'];
+  summaryOptions: SummaryOptions[] = [
+    'Concept',
+    'Context',
+    'Quick',
+    'ParagraphConcept',
+    'ParagraphContext',
+    'RAG',
+    'Vector',
+  ];
   selectedSummaryOption?: SummaryOptions;
 
   explicitUserProfileEnabled: boolean = false; // New property
   viewEnabled: boolean = false;
   ngOnInit() {
-    this.answerSystems = localStorage.getItem('answerSystems') ? JSON.parse(localStorage.getItem('answerSystems') ?? '[]') : [];
+    this.answerSystems = localStorage.getItem('answerSystems')
+      ? JSON.parse(localStorage.getItem('answerSystems') ?? '[]')
+      : [];
 
-    this.answerService.getAnswerSystems().subscribe((systems: System[]) => { 
-      this.answerSystems = systems.filter((system) => system.type !== "conversation");
+    this.answerService.getAnswerSystems().subscribe((systems: System[]) => {
+      this.answerSystems = systems.filter(system => system.type !== 'conversation');
       localStorage.setItem('answerSystems', JSON.stringify(this.answerSystems));
     });
 
-    this.selectedAnswerSystem = localStorage.getItem('selectedSearchAnswerSystem') ?? 'AlbertVector';
-    this.selectedQms = localStorage.getItem('selectedQuerySystem') as QuerySystem ?? 'Dictionary';
-    this.selectedOperator = localStorage.getItem('selectedOperator') as DefaultOperator ?? 'WNEAR';
-    this.selectedSummaryOption = localStorage.getItem('selectedSummaryOption') as SummaryOptions ?? 'Concept';
+    this.selectedAnswerSystem =
+      localStorage.getItem('selectedSearchAnswerSystem') ?? 'AlbertVector';
+    this.selectedQms = (localStorage.getItem('selectedQuerySystem') as QuerySystem) ?? 'Dictionary';
+    this.selectedOperator =
+      (localStorage.getItem('selectedOperator') as DefaultOperator) ?? 'WNEAR';
+    this.selectedSummaryOption =
+      (localStorage.getItem('selectedSummaryOption') as SummaryOptions) ?? 'Concept';
 
     // Retrieve the toggle state from local storage (default to false if not set)
     this.explicitUserProfileEnabled = localStorage.getItem('explicitUserProfileEnabled') === 'true';
@@ -77,7 +99,7 @@ export class SettingsDialogComponent {
   onSystemChange(system: string) {
     localStorage.setItem('selectedSearchAnswerSystem', system);
   }
-  
+
   onBehaviorChange(system: QuerySystem) {
     localStorage.setItem('selectedQuerySystem', system);
   }
@@ -105,9 +127,27 @@ export class SettingsDialogComponent {
       this.settingsService.setPreviewEnabled(inputElement.checked);
     }
   }
-  
 }
 
 export type QuerySystem = 'Dictionary' | 'Index' | 'Answerbank';
-export type DefaultOperator = 'WNEAR' | 'DNEAR' | 'YNEAR' | 'NNEAR' | 'XNEAR' | 'AND' | 'OR' | 'BEFORE' | 'AFTER' | 'SENTENCE' | 'DSENTENCE' | 'PARAGRAPH' 
-export type SummaryOptions = 'Concept' | 'Context' | 'Quick' | 'ParagraphConcept' | 'ParagraphContext' | 'RAG' | 'Vector';
+export type DefaultOperator =
+  | 'WNEAR'
+  | 'DNEAR'
+  | 'YNEAR'
+  | 'NNEAR'
+  | 'XNEAR'
+  | 'AND'
+  | 'OR'
+  | 'BEFORE'
+  | 'AFTER'
+  | 'SENTENCE'
+  | 'DSENTENCE'
+  | 'PARAGRAPH';
+export type SummaryOptions =
+  | 'Concept'
+  | 'Context'
+  | 'Quick'
+  | 'ParagraphConcept'
+  | 'ParagraphContext'
+  | 'RAG'
+  | 'Vector';
