@@ -4,6 +4,6 @@ export const DatabaseActions = createActionGroup({
   source: 'Database',
   events: {
     'Select Databases': props<{ databases: string[] }>(),
-    'Reset Databases': emptyProps()
+    'Reset Databases': emptyProps(),
   },
 });
