@@ -1,17 +1,14 @@
 import { Injectable } from '@angular/core';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 
+import { EMPTY, Observable } from 'rxjs';
 import { concatMap } from 'rxjs/operators';
-import { Observable, EMPTY } from 'rxjs';
 import { DatabaseActions } from './database.actions';
 
 @Injectable()
 export class DatabaseEffects {
-
-
   loadDatabases$ = createEffect(() => {
     return this.actions$.pipe(
-
       ofType(DatabaseActions.loadDatabases),
       /** An EMPTY observable only emits completion. Replace with your own observable API request */
       concatMap(() => EMPTY as Observable<{ type: string }>)
