@@ -1,10 +1,10 @@
 export const environment = {
   production: false,
-  answerserver_api:'api/answerserver',
+  answerserver_api: 'api/answerserver',
   view_api: 'api/view',
   qms_api: 'api/qms',
   dah_api: 'api/dah',
   mlmodels_api: 'api/mlmodels',
   community_api: 'api/community',
-  mistral_api: 'https://llms.idoldemos.net/mistral/v1/completions'
+  mistral_api: 'https://llms.idoldemos.net/mistral/v1/completions',
 };
